Changed vesion of Flyway to 4.0.3
Version of the Flyway was changed to 4.0.3 in order to work
with the utf8mb4 character set in MySQL database.

Changed the size of the unique text fields in some tables,
because the utf8mb4 character set imposes restrictions on
the size of the indexed text fields in InnoDB tables.
For more see:
https://dev.mysql.com/doc/refman/5.5/en/charset-unicode-conversion.html